Replace msgpack-javascript with msgpackr
msgpack-javascript hasn't been published since 5 years ago and
has ~50 weekly downloads from npm (most of these are probably us).

msgpackr was last updated a day ago and has ~200k weekly downloads.

msgpack-javascript is pinned to some very old dependencies including
a version of lodash with known vulnerabilities and it doesn't look
like it's ever going to get touched again. We should abandon it.

msgpackr uses BigInt for the huge numbers in the MessagePack. I
didn't want to mess with any of the logic in toHexDigits() as it
scared me so I thought the safest thing to do was to pass the huge
numbers around as strings.